fixes for design preview

+++ app/Jobs/Util/UnlinkFile.php
@@ -22,15 +22,7 @@ class UnlinkFile implements ShouldQueue
 {
     use Dispatchable, InteractsWithQueue, Queueable, SerializesModels;
 
-    protected $file_path;
-
-    protected $disk;
-
-    public function __construct(string $disk, string $file_path)
-    {
-        $this->file_path = $file_path;
-        $this->disk = $disk;
-    }
+    public function __construct(protected string $disk = '', protected ?string $file_path = ''){}
 
     /**
      * Execute the job.
@@ -40,9 +32,12 @@ class UnlinkFile implements ShouldQueue
     public function handle()
     {
         /* Do not delete files if we are on the sync queue*/
-        if (config('queue.default') == 'sync') {
+        if (config('queue.default') == 'sync') 
             return;
-        }
+        
+
+        if(!$this->file_path)
+            return;
 
         Storage::disk($this->disk)->delete($this->file_path);
     }
